==> notebooks/1. scientometrics_quotes.ipynb <==
import pandas as pd
import numpy as np
import json
import os
from tqdm import tqdm

## Load journal and conferences lists

Set of journal and conferences from [gov.pl](https://www.gov.pl/web/edukacja-i-nauka/nowy-rozszerzony-wykaz-czasopism-naukowych-i-recenzowanych-materialow-z-konferencji-miedzynarodowych) with impact metrics

# load https://www.gov.pl/web/edukacja-i-nauka/nowy-rozszerzony-wykaz-czasopism-naukowych-i-recenzowanych-materialow-z-konferencji-miedzynarodowych
# sheet 1: 'journals'
gov_articles_data = pd.read_excel("..//data//external//Wykaz_dyscyplin_do_czasopism_i_materiałów_konferencyjnych.xlsx", 0, header=None)

gov_articles_data.columns = np.where(gov_articles_data.iloc[0].notna(), 
                                     gov_articles_data.iloc[0].astype(str) + ' - ' + gov_articles_data.iloc[1].astype(str), 
                                     gov_articles_data.iloc[1])
gov_articles_data = gov_articles_data[2:]
gov_articles_data = gov_articles_data.reset_index(drop=True)

gov_articles_data = gov_articles_data.drop('Lp.', axis=1)

display(gov_articles_data)

Get journal titles list

journals_list = gov_articles_data[['Tytuł 2', 'Punkty']] # tytuł 3 - beirzemy tytuł 2 (tłumaczenie tyt 1 jesli było to konieczne) jesli coś jest w tytuł 2 a jesli nie ma nic to bierzemy z tytuł 1
journals_list = journals_list.drop_duplicates(subset=['Tytuł 2'])
journals_list = journals_list.dropna()
journals_list = journals_list.set_index('Tytuł 2')
display(journals_list)

# load https://www.gov.pl/web/edukacja-i-nauka/nowy-rozszerzony-wykaz-czasopism-naukowych-i-recenzowanych-materialow-z-konferencji-miedzynarodowych
# sheet 2: 'conferences'
gov_conferences_data = pd.read_excel("..//data//external//Wykaz_dyscyplin_do_czasopism_i_materiałów_konferencyjnych.xlsx", 1)
gov_conferences_data = gov_conferences_data.drop('LP.', axis=1)

gov_conferences_data['Przypisane dyscypliny naukowe'] = gov_conferences_data['Przypisane dyscypliny naukowe'].replace('\n', ' ', regex=True)

display(gov_conferences_data)

Get conferences titles list


conferences_list = gov_conferences_data[['Nazwa konferencji', 'Liczba punktów']]
conferences_list = conferences_list.drop_duplicates(subset=['Nazwa konferencji'])
conferences_list = conferences_list.dropna()
conferences_list = conferences_list.set_index('Nazwa konferencji')
display(conferences_list)

## Load collection of computer science publications

A collection of data on publications in the field of computer science from [dblp](https://paperswithcode.com/dataset/dblp)

# load https://www.aminer.org/citation v10
directory_path = "../data/external/dblp-ref-10"

publications = []

try:
    file_names = os.listdir(directory_path)

    # Wyświetl nazwy plików
    for file_name in file_names:
        print(file_name)
        with open(os.path.join(directory_path, file_name)) as input_file:
            for line in tqdm(input_file):
                publication = json.loads(line)
                if 'references' not in publication:
                    publication['references'] = []
                if 'authors' not in publication:
                    publication['authors'] = []
                publications.append(publication)
        
except FileNotFoundError:
    print(f"Katalog '{directory_path}' nie istnieje.")

except PermissionError:
    print(f"Brak uprawnień do odczytu katalogu '{directory_path}'.")

display(len(publications))

## Concatenate datasets

First, convert the list to a dictionary; this allows us to search by title using a hash with **O(1)** complexity.

# convert articles list
journal_points_lookup = journals_list.to_dict(orient='index')

# convert conferences list
conference_points_lookup = conferences_list.to_dict(orient='index')

Next find matching venues & titles

processed_publications = []

for publication in tqdm(publications):
    if publication['venue'] not in conference_points_lookup and publication['venue'] not in journal_points_lookup:
        continue

    processed_publication = {key: publication[key] for key in ["id", "title", "year", "references", "authors", "n_citation", "venue"]}
    if publication['venue'] in conference_points_lookup:
        processed_publication['gov_score'] = conference_points_lookup[publication['venue']]['Liczba punktów']
    elif publication['venue'] in journal_points_lookup:
        processed_publication['gov_score'] = journal_points_lookup[publication['venue']]['Punkty']
    else:
        continue

    processed_publications.append(processed_publication)

del publications

print(len(processed_publications))

**Note**
When we filter by *venue* in both list program match the vast majority of records than when we use *title* for publications and *venue* for conferences separately (about 18 rows)

## Save filtered data

tqdm.pandas()

processed_publications = pd.DataFrame(processed_publications)

processed_publications.progress_apply(
    lambda row: row,
    axis=1
).to_csv(
    os.path.join("../data/interim", "articles_with_score_df.csv"),
    index=False
)
