==> notebooks/3. scientometrics_queries_generator.ipynb <==
import pandas as pd
import numpy as np
import os
import random
from tqdm import tqdm

## Load data

Queries generator will use set of four lists with static words:
1. Nouns
2. Verbs
3. Adjectives
4. Gerund (participles)

collections = ["nouns", "verbs", "adjectives", "participles", "gerounds"]
dfs = {}

for name in collections:
    column_name = name.capitalize()
    csv_path = f"../data/raw/{name}.csv"
    try:
        df = pd.read_csv(csv_path, header=None, names=[column_name])
        dfs[name] = df
    except FileNotFoundError:
        print(f"Nie można znaleźć pliku {csv_path}, pomijam...")
    except Exception as e:
        print(f"Wystąpił błąd podczas odczytu pliku {csv_path}: {e}")

nouns_df = dfs["nouns"]
verbs_df = dfs["verbs"]
adjectives_df = dfs["adjectives"]
participles_df = dfs["participles"]

## Query generator

Sample code snippet to generate sentences that are candidate phrases for search queries. Which will be used in a search engine to find articles on a similar topic.

def query_generator(nouns, verbs, adjectives, participles, limit):
    generated_queries = set()
    
    while len(generated_queries) < limit:
        noun = random.choice(nouns)[0]
        verb = random.choice(verbs)[0]
        adjective = random.choice(adjectives)[0]
        participle = random.choice(participles)[0]
        
        query = f"{noun} {verb} {adjective} {participle}"
        
        if query not in generated_queries:
            generated_queries.add(query)
            yield query

### 1. Language Tool

First attempt to make queries more natrual with `Language Tool` library

Source: [Githube](https://github.com/Findus23/pyLanguagetool) </br>
Tutorial: [here](https://www.kaggle.com/code/yeoyunsianggeremie/how-to-use-language-tool-python-without-internet)

from language_tool_python import LanguageTool

def correct_sentence_lt(sentence: list, debug: bool = False) -> str:
    """
    Corrects a sentence using LanguageTool.

    Args:
        sentence (str): The sentence to correct.
        debug (bool, optional): A flag indicating whether to display matches, defaults to False.

    Returns:
        str: The corrected sentence.
    """
    lt = LanguageTool('en-US')
    matches = lt.check(sentence)

    if debug:
        display(matches)

    return lt.correct(sentence)

### 2. Gramformer

Second attempt to build more consistent sentences this time with `Gramformer`

Source: [Githube](https://github.com/thevkrant/gramformer) </br>
Tutorial: [here](https://www.vennify.ai/gramformer-correct-grammar-transformer-nlp/)

from gramformer import Gramformer
import torch

def set_seed(seed):
    torch.manual_seed(seed)
    if torch.cuda.is_available():
        torch.cuda.manual_seed_all(seed)
    else:
        print(f"`Cuda` is unavailable")

def correct_sentence_gf(gf: Gramformer, sentence: str, debug = False) -> str:
    """
    Corrects a sentence using Gramformer.

    Args:
        gf (Gramformer): The Gramformer object to use for correction.
        sentence (str): The sentence to correct.
        debug (bool, optional): A flag indicating whether to display matches, defaults to False.

    Returns:
        str: The corrected sentence.
    """

    return gf.correct(sentence, max_candidates=1)[0]

In this section, we tested two libraries for automatic sentence correction: LanguageTool and Gramformer. The first tool (LanguageTool) did not perform very well and often failed to improve the sentences meaningfully. In contrast, Gramformer generally produced better and more fluent sentence corrections.

However, the final decision was to retain the raw version.

## Prepare queries

queries = query_generator(
    nouns_df.values.tolist(), 
    verbs_df.values.tolist(), 
    adjectives_df.values.tolist(), 
    participles_df.values.tolist(), 
    850000
)

queries_df = pd.DataFrame({'query': queries})

queries_df.drop_duplicates(subset=['query'])

unique_queries_count = len(queries_df['query'].unique())
print(f"Obtained {unique_queries_count} unique queries.")

### Generate embeddings for queries

We generate embeddings for raw queries to expedite the target experiment. By querying the vector database ChromaDB with pre-embedded content, we offload the burden of embedding the raw query received in the request before searching and returning the K most similar titles of scientific papers.

from tqdm import tqdm
from sentence_transformers import SentenceTransformer, util
import torch

if torch.cuda.is_available():
    print("CUDA is available: ", torch.cuda.is_available())   
    print("Number of CUDA devices: ", torch.cuda.device_count())
    print("CUDA current device: ", torch.cuda.current_device())
    print("CUDA device name: ", torch.cuda.get_device_name(0))

    model = SentenceTransformer("all-MiniLM-L6-v2", device='cuda')
else:
    model = SentenceTransformer("all-MiniLM-L6-v2", device='cpu')

titles = queries_df['query'].tolist()
batch_size = 2000

titles_embeddings = []

for i in tqdm(range(0, len(titles), batch_size), desc="Embedding titles"):
    batch = titles[i:i + batch_size]
    batch_embeddings = model.encode(batch)
    titles_embeddings.extend(batch_embeddings)

queries_df['embedding'] = titles_embeddings

queries_df.to_pickle('..//data//interim//queries_with_embeddings.pkl')
